Rename control-story-shared to shared
